added buildHeader function
This programmatically generates the header and nav for any given page. I've made this in the templates script so it can be standardized and iterated upon much more easily across the entire site without having to manually update every single page.

The nav is built out of the provided file pathname and the theme button is replacing the previous color palette button.