footer logo
Since footer is procedurally generated, I have to primary logos and that's just too much. I either have to modify the generation to make the logo modular or I can just style it specifically on the about page. I'm choosing the latter for now, letting the footer version be just an outline logo.