properly handle the blank dropdown option
When the blank default option was selected, it used to run into an error because it would try to generate a report that matched its id but it wouldn't find any. Now it clears any visible reports no matter what dropdown choice is selected, but then exits the function early if it's the default option.