restructured and restyled support links
Now a and strong tags are nested inside a custom support tag so I can decouple the animated element from the clickable element. It should no longer be possible to hover your cursor on the edge of a moving element, having it flash rapidly. It only occurred in a small area, but now it shouldn't happen at all and it should be an easier target to click on. It's a bit cleaner.

I changed the animation to animate both up and down from the base transform so it's better vertically centered.

On hover it no longer gets a background color. Also the glow is tighter and text / glow colors have been updated.