adjusted playlist button behavior
Now the clickable elements are detached from the elements that detect a hover, just like I did with the support links. This prevents the issue where hovering on the edge of the element would cause quick flashing as the button animates away from the cursor, escaping the cursor, dropping back down, and reactivating.