rewards button behavior / styling
The colored outline that indicates reward type is now made via css rather than baked into the image. Hovering over the image now changes the border color to white.

I also decoupled the hover target to an element dedicated to that so I can eliminate the flashing bug that happens when you hover over the far edge of the link.

I also also updated some styling and arrangement of information.